rules: move rules*.go contents to manager*.go
