+++ rules/manager_test.go
